Weapon edits
created pickup system for weapons and future health packs and potentially ammo.

noticed issue with muzzleflash and HitByRay functions - only work with one game object, function needs to be changed to allow for multiple different GameObjects with the same script attached.

implemented a cooldown for weapons

added a change in range based on which weapon - pistols shouldn't shoot as far as rifles.